backend/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
djoser
psycopg2-binary==2.8.6
python-dotenv==0.21.0
Pillow==7.0.0
django-filter==2.4.0
